refactor(tree-sitter): split module into multiple files
